Refactor code into printer, statistics, tests and utils
