+++ Cargo.toml
@@ -20,3 +20,4 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1"
 futures = "0.3.21"
 clap = "3.1.8"
+csv = "1.1.6"
